refactor: split remaining preferences sections (#119)
Closes #119
